*refactor : 라인 생성 분리 (Line -> Points)
